precision updates

+++ model/domain.go
@@ -5,27 +5,27 @@ import "time"
 type (
 	SatelliteData struct {
 		TempSatelliteID string    `json:"name" gorm:"primaryKey"`
-		Distance        float32   `json:"distance"`
+		Distance        float64   `json:"distance"`
 		Message         string    `json:"message"`
 		CreatedAt       time.Time `json:"created_at" gorm:"autoUpdateTime"`
 	}
 
 	TempSatellite struct {
 		ID            string         `json:"name" gorm:"primaryKey"`
-		X             float32        `json:"x"`
-		Y             float32        `json:"y"`
+		X             float64        `json:"x"`
+		Y             float64        `json:"y"`
 		CreatedAt     time.Time      `json:"created_at" gorm:"autoUpdateTime"`
 		SatelliteData *SatelliteData `json:"satellite_data" gorm:"foreignKey:TempSatelliteID"`
 	}
 
 	Satellite struct {
 		Name     string   `json:"name" validate:"notnull,notblank"`
-		Distance float32  `json:"distance" validate:"notnull"`
+		Distance float64  `json:"distance" validate:"notnull"`
 		Message  []string `json:"message" validate:"notnull,nonempty"`
 	}
 
 	ShipDataResponse struct {
-		Position map[string]float32 `json:"position"`
+		Position map[string]float64 `json:"position"`
 		Message  string             `json:"message"`
 	}
 )
